requirements/base.txt
django-allauth
djangorestframework
django-filter
dj_database_url
psycopg2-binary
python-decouple
django-versatileimagefield